refactor: 카드 정보 context의 provider 분리 및 사용 범위 확대
